requirements.txt
flask
PyPDF2
langchain
faiss-cpu
requests
numpy
python-dotenv
Markdown
openai
celery
redis